requirements.txt
google-cloud-aiplatform
fastavro
avro
apache_beam
flask
sqlalchemy
jwt
joblib
sklearn


